gdb_spawn_attach_cmdline: use unsupported instead of untested
In a previous commit (899fa3f87d3: gdb/testsuite: Introduce and use
gdb_spawn_attach_cmdline), if gdb_spawn_attach_cmdline cannot have GDB
attach to the process because of ptrace restrictions (operation not
permitted), the proc issues UNTESTED.  This should really be
UNSUPPORTED, as it is done in gdb_attach.

This patch fixes this oversight.

Change-Id: Ib87e33b9230f3fa7a85e06220ef4c63814b71f7d